refactor existing example
