+++ src/utils/mobxSearchTree.js
@@ -8,6 +8,7 @@ export default (tree, service) =>
       snapshot: toJS,
       extend: extendObservable,
       service,
+      debounce: 500
     },
     tree
   )
